backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI-powered document processing system with human review workflow"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "celery[redis]>=5.6.2",
    "fastapi>=0.128.8",
    "google-genai>=1.0.0",
    "httpx>=0.28.1",
    "pandas>=3.0.0",
    "prometheus-client>=0.24.1",
    "pyarrow>=23.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "testing.postgresql>=1.3.0; sys_platform != 'win32'",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
